update master from RIOT-OS repository

+++ boards/feather-nrf52840/include/periph_conf.h
@@ -39,8 +39,8 @@ static const uart_conf_t uart_config[] = {
         .dev        = NRF_UARTE0,
         .rx_pin     = GPIO_PIN(0,24),
         .tx_pin     = GPIO_PIN(0,25),
-        .rts_pin    = (uint8_t)GPIO_UNDEF,
-        .cts_pin    = (uint8_t)GPIO_UNDEF,
+        .rts_pin    = GPIO_UNDEF,
+        .cts_pin    = GPIO_UNDEF,
         .irqn       = UARTE0_UART0_IRQn,
     },
 };
